Fixed build errors. Needs reviewing.
Build errors that have occurred due to the merge with master have been fixed but a review is needed to make sure existing functionality have not been overwritten. Tests have not be run yet either. Checking this in as significant fixes were added.

+++ src/SFA.DAS.EmployerApprenticeshipsService.Domain/Interfaces/IApprenticeshipInfoServiceWrapper.cs
@@ -8,6 +8,6 @@ namespace SFA.DAS.EAS.Domain.Interfaces
     {
         Task<StandardsView> GetStandardsAsync(bool refreshCache = false);
         Task<FrameworksView> GetFrameworksAsync(bool refreshCache = false);
-        ProvidersView GetProvider(int ukPrn);
+        ProvidersView GetProvider(long ukPrn);
     }
 }
